Set undoManager on a few contexts to nil.
Also set a merge policy for the event store context. It should not be needed, but Kevin Hoctor is getting conflicts between PSC and Store global ids on 10.9. This should not be possible, because there is only one MOC accessing the store, but it seems to happen when a save interrupts a merge.

This issue has been reported for 10.9 online, and seems to be a genuine Apple bug (http://lists.apple.com/archives/cocoa-dev/2005/Aug/msg00899.html).