Update Ensembles.podspec
Set ios.deployment_target = '7.0' and osx.deployment_target = '10.10' on the `Multipeer` subspace, not on the main spec.  This was forcing iOS 7 or OS X 10.10, while the primary spec requires only iOS 6.0 or OS X 10.7.  (The README suggests iOS 7 and OS X 10.9 are the desired minimums; this change still applies, but the targets should be set if needed in lines 24 & 25).